Add ship color selection Need to add velocity and power
