+++ src/seagulls/examples/space_shooter/_selectable_ship_menu.py
@@ -19,7 +19,6 @@ from seagulls.engine import (
     SurfaceRenderer
 )
 from seagulls.examples import ISetActiveScene
-from seagulls.examples.space_shooter._active_ship_client import ISetActiveShip
 
 logger = logging.getLogger(__name__)
 
@@ -80,6 +79,34 @@ class BlueShip(IShip):
 
     def offset(self) -> int:
         return 512
+
+
+class IProvideActiveShip(ABC):
+
+    @abstractmethod
+    def get_active_ship(self) -> IShip:
+        pass
+
+
+class ISetActiveShip(ABC):
+
+    @abstractmethod
+    def set_active_ship(self, ship: IShip) -> None:
+        pass
+
+
+class ActiveShipClient(IProvideActiveShip, ISetActiveShip):
+
+    _active_ship: IShip
+
+    def __init__(self, ship: IShip):
+        self._active_ship = ship
+
+    def get_active_ship(self) -> IShip:
+        return self._active_ship
+
+    def set_active_ship(self, ship: IShip) -> None:
+        self._active_ship = ship
 
 
 class ShipButton(GameObject):
